hide pools without a club linking to it (#94)
With this change pools referred to by no clubs (with a link such as `[[<pool short name>]]` in the club markdown) no longer appear to general page viewers or on the map. When logged into an account with edit access the pools appear on the place page, but still don't appear on the map. This reduces clutter of unused pools. Long term I'd like something like https://github.com/TomGoBravo/tourist-with-flask/issues/92 to hold a large repository of pools separate from hockey tourist data.